Implement Rue Runtime Refactor - Phase 1 and 2 complete
**Phase 1 - Structural Split (No Behavior Changes) ✅**
- Create module files in crates/rue-codegen/src/runtime/:
  - mod.rs: Re-exports & glue
  - context.rs: RuntimeContext struct + helper methods  
  - startup.rs: _start, __rue_main, signal setup
  - data.rs: Static data section (strings, buffers)
  - syscalls.rs: Syscall wrappers and helpers
  - io.rs: println variants, input
  - conversion.rs: itoa, atoi, to_i32, to_i64
  - x86_64.rs: generate_runtime() orchestrator only

**Phase 2 - Minimal Syscall API Layer ✅**
- Implement small, direct syscall wrappers in syscalls.rs:
  - sys_write(fd, buf_reg, len_reg)
  - sys_read(fd, buf_reg, len) 
  - sys_exit(code_reg)
- Replace inline syscall logic in some runtime helpers
- Keep emit_syscall_with_error_check for compatibility

**Phase 3 - Data Section Groundwork ✅**
- Add placeholder .bss section symbol for heap allocation
- Document where heap metadata will live with RUE_HEAP_START constant

**Testing & Validation ✅**
- All corpus tests pass
- All unit tests pass (buck2 test //crates/...)
- Generated executables work identically to pre-refactor
- No changes to generated assembly behavior

✅ **Expected Outcome Achieved:**
- generate_runtime() is now a clean orchestrator
- Syscall logic is centralized with minimal API layer
- Heap groundwork exists for future bump allocator
- Identical runtime behavior to pre-refactor
- Ready for easy addition of mmap/brk and memory helpers
