+++ crates/rue-codegen/src/runtime/mod.rs
@@ -4,6 +4,12 @@
 //! that are embedded in Rue executables. It provides assembly code for basic I/O
 //! operations, syscalls, and other runtime support functions.
 
+pub mod context;
+pub mod conversion;
+pub mod data;
+pub mod io;
+pub mod startup;
+pub mod syscalls;
 pub mod x86_64;
 
 pub use x86_64::generate_runtime;
